backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic[email]
motor>=2.5.0
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart 